lib/cdda_interface/toc.c: fixed pregap and first track handling.
It was broken in commit c94fcc7b04... which removed special handling of track 0 (representing the pregap/hidden track one) always returning error -401 in that case. Instead, the pregap code was moved to a block handling the first track (track 1).

Testing can be further improved by adding tests on images with pregap.