Adds --force-overread
Force overreading into the lead-out portion of the disc. This option is
only applicable when using the-O option with a positive sample offset
value. Many drives are not capable of reading into this portion of the
disc and attempting to do so on those drives will produce read errors
and possibly hard lockups.

See:

https://gist.github.com/AnwarShah/560d77f7d7da52553f918f3ecb7401e3