configure.ac: fix out-of-tree tests
To reproduce the test failure one needs to run
configure using absolute path in a directory
outside source tree. For example:

    $ $(pwd)/../libcdio-paranoia/configure
    $ make
    $ make check

This will cause 'native_abs_top_srcdir' to contain wrong path.

It happens because '[]' is an escape in autoconf.
As a the following configure.ac snippet:
    [\\/]* | ?:[\\/]* )  # Absolute name.
gets translated into the following shell code:
    \\/* | ?:\\/* )  # Absolute name.

The fix is to change quotes from '[]' for a short while.

Reported-by: eroen
Reported-by: Paolo Pedroni
Bug: https://bugs.gentoo.org/546388
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>