Require libcdio at least 2.0.0
I doubt 0.83 still works, although I haven't check (and don't have much
interest in checking).

A *lot* has gone on since 0.83 so even if it still works, strong-arming
folks to get up to 2.0.0 is probably a good thing (from the developer
side if not the end-user side) in the long run.

And it isn't like 2.0.0 from the last day of 2017 isn't is all *that* new.